Fix lsp issues when opening Syft repository
Test fixtures for this project have some symlink loops on purpose
to verify that Syft handles them correctly when indexing the file system.

These seem to create issues for lsp as it gets into a tight loop when
watching directories when I open the repository, so ignore the test-fixtures
dir.